+++ packages/waku/src/lib/vite-entries/entry.browser.tsx
@@ -1,2 +1,2 @@
-import './lib/browser-preamble.js';
+import '../vite-rsc/browser-preamble.js';
 import 'virtual:vite-rsc-waku/client-entry';
