+++ packages/waku/src/lib/vite-entries/entry.ssr.tsx
@@ -1,0 +1,1 @@
+export { renderHTML, renderHtmlFallback } from '../vite-rsc/ssr.js';
