+++ packages/waku/src/lib/vite-rsc/deploy/aws-lambda/entry.ts
@@ -1,6 +1,6 @@
 import { Hono } from 'hono';
 import * as honoAwsLambda from 'hono/aws-lambda';
-import { createHonoHandler } from '../../lib/engine.js';
+import { createHonoHandler } from '../../engine.js';
 import { honoEnhancer } from 'virtual:vite-rsc-waku/hono-enhancer';
 import { config } from 'virtual:vite-rsc-waku/config';
 import { serveStatic } from '@hono/node-server/serve-static';
@@ -29,4 +29,4 @@ export const handler: any = import.meta.env.WAKU_AWS_LAMBDA_STREAMING
   ? honoAwsLambda.streamHandle(app)
   : honoAwsLambda.handle(app);
 
-export { handleBuild } from '../../lib/build.js';
+export { handleBuild } from '../../build.js';
