+++ packages/waku/src/lib/vite-rsc/deploy/netlify/entry.ts
@@ -1,5 +1,5 @@
 import { Hono } from 'hono';
-import { createHonoHandler } from '../../lib/engine.js';
+import { createHonoHandler } from '../../engine.js';
 import { honoEnhancer } from 'virtual:vite-rsc-waku/hono-enhancer';
 import { INTERNAL_setAllEnv } from '../../../../server.js';
 
@@ -21,4 +21,4 @@ const app = honoEnhancer(createApp)(new Hono());
 export default async (request: Request, context: unknown) =>
   app.fetch(request, { context });
 
-export { handleBuild } from '../../lib/build.js';
+export { handleBuild } from '../../build.js';
