+++ packages/waku/src/lib/vite-rsc/deploy/vercel/entry.ts
@@ -1,6 +1,6 @@
 import { getRequestListener } from '@hono/node-server';
 import { Hono } from 'hono';
-import { createHonoHandler } from '../../lib/engine.js';
+import { createHonoHandler } from '../../engine.js';
 import { honoEnhancer } from 'virtual:vite-rsc-waku/hono-enhancer';
 import { config } from 'virtual:vite-rsc-waku/config';
 import path from 'node:path';
@@ -25,4 +25,4 @@ const app = honoEnhancer(createApp)(new Hono());
 
 export default getRequestListener(app.fetch);
 
-export { handleBuild } from '../../lib/build.js';
+export { handleBuild } from '../../build.js';
