--- packages/waku/src/lib/vite-rsc/entry.ssr.tsx
@@ -1,1 +1,0 @@
-export { renderHTML, renderHtmlFallback } from './lib/ssr.js';
