+++ packages/waku/src/lib/vite-types.d.ts
@@ -2,7 +2,7 @@
 /// <reference types="@vitejs/plugin-rsc/types" />
 
 declare module 'virtual:vite-rsc-waku/server-entry' {
-  const default_: import('../types.ts').EntriesDev['default'];
+  const default_: import('./types.ts').EntriesDev['default'];
   export default default_;
 }
 
@@ -15,16 +15,16 @@ declare module 'react-dom/server.edge' {
 declare module 'virtual:vite-rsc-waku/set-platform-data' {}
 
 declare module 'virtual:vite-rsc-waku/middlewares' {
-  export const middlewares: import('../../config.ts').Middleware[];
+  export const middlewares: import('../config.ts').Middleware[];
 }
 
 declare module 'virtual:vite-rsc-waku/hono-enhancer' {
-  export const honoEnhancer: import('../../cli.ts').HonoEnhancer;
+  export const honoEnhancer: import('../cli.ts').HonoEnhancer;
 }
 
 declare module 'virtual:vite-rsc-waku/config' {
-  export const flags: import('./plugin.ts').Flags;
-  export const config: Required<import('../config.ts').Config>;
+  export const flags: import('./vite-rsc/plugin.ts').Flags;
+  export const config: Required<import('../config/types.ts').Config>;
   export const isBuild: boolean;
 }
